Rename of files and progress from last four weeks
I renamed Storage.hs to S.hs and kept progress not visible due to, well, some superstitiousness of me.

I prefer to make file names as short as possible because it seems I better work with them in hard cases. So I renamed Storage.hs. Again, sometimes I use public progress information as a stimulae, sometimes it hinders my work. The latter happen when I work sporadically, as with this project.

As I have most of data structures and tests in place, I can now put progress online. So be it.